split hash_table.rs into small files
